DRFAuth/requirements.txt
Django
djangorestframework
psycopg2-binary
PyJWT
bcrypt
python-dotenv
django-environ